refactored SCR processing, removed use of __CONTRIBUTORS__, code cleanup, #244
Previously a single variable containing a combiner header, footer, and __CONTRIBUTORS__-tag
was generated, and then the __CONTRIBUTORS__-tag used/replaced with the
names of those that contributed.
This has since been removed.

Now, a HEADER and seperate FOOTER variable are created and the CONTRIBUTORS
is just added after the HEADER and the FOOTER follows.
This simplifies the code and makes the SCR process more similar to the BIN process.

Also identified that the "*.sh" were previously considered as TXT, but are SCR-ipts.

Also changed $x to be "${x}" for consistency (for space in filename)

The SCR-processing now adds to git the files modified, ready for verification and commiting.