cleanup of BIN and SCR processing, formatting, #244
moved *.MD from TXT to BIN (where a file.license is created)
as adding comments to these files was difficult.
Refer:
 https://stackoverflow.com/questions/4823468/comments-in-markdown
 https://stackoverflow.com/questions/37364722/how-do-i-add-inline-comments-in-github-flavored-markdown

added *.V, *.VH and *.S to to TXT-file
 These were previously skipped due to being matched for VHD, VHD and ASM respectively

added 'space'-char around the file-extension to match whole-words,
 due to "v" being seen as a "vhdl"

added a check to see if the existing file contains copyright/license info,
if so, report this as file.5.*

added support for dry-run by using git: add/diff/reset/checkout

added summary/warning at the end of the script

removed files of type "cfg" "ucf" "xise" as these are not supported.